requirements.txt
astroid==3.3.9
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dill==0.4.0
dominate==2.9.1
flake8==7.2.0
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
pylint==3.3.7
python-dotenv==1.1.0
tomlkit==0.13.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
